requirements.txt
Flask
scikit-learn
pandas
numpy
qrcode[pil]
pyzbar
Pillow
opencv-python-headless
joblib
